toolips docs

a manic web-development framework

Toolips.jl is a reactive. full-stack, modular and extendable web-development framework written in pure Julia.

core

requests

connections

routing

servers

server extensions

developer api

creating servable extensions

creating server extensions

toolips internals

projects

creating a project

project tutorials

a repl crash course

project walkthrough

deploying a toolips server

servables

components

style components

file

toolips session extension

tutorials

Basics - Creating an API

Basics - Building a static site

Intermediate - Building Your First WebApp

Advanced - Creating a ServerExtension

curated projects

Below are some curated and deployed toolips apps and toolips extensions that can be used as examples or loaded as a server extension to look into.

modules

sites

EmsComputer.jl - ems computer

ToolipsApp.jl - toolips app

server extensions

ToolipsModifier.jl - Reactive Components.

ToolipsRemote.jl - Allows remote management of a Toolips server via a

regular HTTP request.

ToolipsDefaults.jl - More style components, and style default methods.

servables

ToolipsDefaults.jl - More style components, and style default methods.

ToolipsCanvas.jl - Interact with the Canvas element via script.

Toolips — Module

Created in February, 2022 by chifi - an open source software dynasty. by team toolips This software is MIT-licensed.

Toolips

Toolips.jl is a fast, asynchronous, low-memory, full-stack, and reactive web-development framework always written in pure Julia.

Module Composition

Toolips