creating a project

Toolips projects are created using the newapp and newwebapp methods respectively. newapp will create a simple project and newwebapp will create a full-stack web-app.

micro

Toolips.new_app — Function

new_app(::String) -> _

Creates a minimalistic app, usually used for creating endpoints – but can be used for anything. For an app with a real front-end, it might make sense to add some extensions.

example

full-stack

Toolips.new_webapp — Function

new_webapp(::String) -> _

Creates a fully-featured web-app. Adds CanonicalToolips.jl to provide more high-level interface origrannubg from Julia.

example