servables

Servables are any non-core data-structure that is built with the objective of being written to a stream.

Toolips.Servable — Type

abstract type Servable

Servables can be written to a Connection via thier f() function and the interface. They can also be indexed with strings or symbols to change properties

Consistencies

f::Function - Function whose output to be written to http().

components

A component is a Servable which contains markup information and can easily be translated into elements with properties..

Toolips.Component — Type

Component <: Servable

name::String f::Function properties::Dict –––––––––

name::String - The name field is the way that a component is denoted in code.

f::Function - The function that gets called with the Connection as an

argument.

properties::Dict - A dictionary of symbols and values.

constructors

Component(name::String, tag::String, properties::Dict)

Indexing a component will yield its .properties:

Base.getindex — Method

Interface

getindex(::Servable, ::Symbol) -> ::Any

Returns a property value by symbol or name.

example

Base.getindex — Method

Interface

getindex(::Servable, ::String) -> ::Any

Returns a property value by string or name.

example

Base.setindex! — Method

Interface

setindex!(::Servable, ::Symbol, ::Any) -> ::Any

Sets the property represented by the symbol to the provided value.

example

Missing docstring.

Missing docstring for setindex!(::Servable, ::Any, ::Symbol). Check Documenter's build log for details.

Base.setindex! — Method

Interface

setindex!(::Servable, ::String, ::Any) -> ::Any

Sets the property represented by the string to the provided value.

example

There is a library of default components that comes with toolips. Generally, their name coincides with a docstring. All of these take an infinite number of key-word arguments. These arguments become the properties of a Servable.

Missing docstring.

Missing docstring for img. Check Documenter's build log for details.

Missing docstring.

Missing docstring for link. Check Documenter's build log for details.

Missing docstring.

Missing docstring for meta. Check Documenter's build log for details.

Missing docstring.

Missing docstring for input. Check Documenter's build log for details.

Missing docstring.

Missing docstring for a. Check Documenter's build log for details.

Missing docstring.

Missing docstring for p. Check Documenter's build log for details.

Missing docstring.

Missing docstring for h. Check Documenter's build log for details.

Missing docstring.

Missing docstring for img. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ul. Check Documenter's build log for details.

Missing docstring.

Missing docstring for li. Check Documenter's build log for details.

Missing docstring.

Missing docstring for divider. Check Documenter's build log for details.

Missing docstring.

Missing docstring for br. Check Documenter's build log for details.

Missing docstring.

Missing docstring for i. Check Documenter's build log for details.

Missing docstring.

Missing docstring for title. Check Documenter's build log for details.

Missing docstring.

Missing docstring for span. Check Documenter's build log for details.

Missing docstring.

Missing docstring for iframe. Check Documenter's build log for details.

Missing docstring.

Missing docstring for svg. Check Documenter's build log for details.

Missing docstring.

Missing docstring for element. Check Documenter's build log for details.

Missing docstring.

Missing docstring for label. Check Documenter's build log for details.

Missing docstring.

Missing docstring for script. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nav. Check Documenter's build log for details.

Missing docstring.

Missing docstring for button. Check Documenter's build log for details.

Missing docstring.

Missing docstring for form. Check Documenter's build log for details.

We can also compose components together using push! and style them using style!

Missing docstring.

Missing docstring for push!(::Servable, ::Servable). Check Documenter's build log for details.

Missing docstring.

Missing docstring for push!(::Servable, ::Vector{Servable}). Check Documenter's build log for details.

Base.push! — Method

Interface

push!(::Component, ::Component ...) -> ::Component

example

Toolips.properties! — Method

Interface

properties!(::Servable, ::Servable) -> _

Copies properties from s,properties into c.properties.

example

Base.push! — Method

Interface

push!(::Component, ::Component) ->

example

Toolips.style! — Function

Interface

style!(::Servable, ::Style) -> _

Applies the style to a servable.

example

Interface

style!(::Style, ::Style) -> _

Copies the properties from the second style into the first style.

example

style components

Style components are change the style of a Component

Missing docstring.

Missing docstring for StyleComponent. Check Documenter's build log for details.

The main style components are Animations and Styles.

Missing docstring.

Missing docstring for Animation. Check Documenter's build log for details.

Animating and property adjustment is done with indexing.

file

The file Servable, as you might expect, serves a file via a directory.

Toolips.File — Type

File <: Servable

dir::String f::Function –––––––––

field info

dir::String - The directory of a file to serve.

f::Function - Function whose output to be written to http().

constructors

File(dir::String)