servables

Servables are any non-core data-structure that is built with the objective of being written to a stream.

Toolips.Servable — Type

abstract type Servable

Servables can be written to a Connection via thier f() function and the interface. They can also be indexed with strings or symbols to change properties

Consistencies

f::Function - Function whose output to be written to http().

properties::Dict - The properties of a given Servable. These are written

into the servable on the calling of f().

components

A component is a Servable which contains markup information and can easily be translated into elements with properties..

Toolips.Component — Type

Component <: Servable

name::String f::Function properties::Dict –––––––––

name::String - The name field is the way that a component is denoted in code.

f::Function - The function that gets called with the Connection as an

argument.

properties::Dict - A dictionary of symbols and values.

constructors

Component(name::String, tag::String, properties::Dict)

There is a library of default components that comes with toolips. Generally, their name coincides with a docstring. All of these take an infinite number of key-word arguments. These arguments become the properties of a Servable.

Missing docstring.

Missing docstring for img. Check Documenter's build log for details.

Missing docstring.

Missing docstring for link. Check Documenter's build log for details.

Missing docstring.

Missing docstring for meta. Check Documenter's build log for details.

Missing docstring.

Missing docstring for input. Check Documenter's build log for details.

Missing docstring.

Missing docstring for a. Check Documenter's build log for details.

Missing docstring.

Missing docstring for p. Check Documenter's build log for details.

Missing docstring.

Missing docstring for h. Check Documenter's build log for details.

Missing docstring.

Missing docstring for img. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ul. Check Documenter's build log for details.

Missing docstring.

Missing docstring for li. Check Documenter's build log for details.

Missing docstring.

Missing docstring for divider. Check Documenter's build log for details.

Missing docstring.

Missing docstring for br. Check Documenter's build log for details.

Missing docstring.

Missing docstring for i. Check Documenter's build log for details.

Missing docstring.

Missing docstring for title. Check Documenter's build log for details.

Missing docstring.

Missing docstring for span. Check Documenter's build log for details.

Missing docstring.

Missing docstring for iframe. Check Documenter's build log for details.

Missing docstring.

Missing docstring for svg. Check Documenter's build log for details.

Missing docstring.

Missing docstring for element. Check Documenter's build log for details.

Missing docstring.

Missing docstring for label. Check Documenter's build log for details.

Missing docstring.

Missing docstring for script. Check Documenter's build log for details.

Missing docstring.

Missing docstring for nav. Check Documenter's build log for details.

Missing docstring.

Missing docstring for button. Check Documenter's build log for details.

Missing docstring.

Missing docstring for form. Check Documenter's build log for details.

style components

Style components are change the style of a Component

files