advanced usage

Welcome to the Toolips developer API. This section describes making toolips extensions as well as

creating servables

Servables are probably the most approachable type to make for your first extension. Servable extensions work by simply making a sub-type of Servable. For example, the Component's source code:

function Component(name::String = "", tag::String = "",
     properties::Dict = Dict{Any, Any}())
     push!(properties, :children => Vector{Servable}())
     extras = Vector{Servable}()
     f(c::AbstractConnection) = begin
         open_tag::String = "<$tag id = $name "
         text::String = ""
         write!(c, open_tag)
         for property in keys(properties)
             special_keys = [:text, :children]
             if ~(property in special_keys)
                 prop::String = string(properties[property])
                 propkey::String = string(property)
                 write!(c, " $propkey = $prop ")
             else
                 if property == :text
                     text = properties[property]
                 end
             end
         end
         write!(c, ">")
         if length(properties[:children]) > 0
             write!(c, properties[:children])
        end
        write!(c, "$text</$tag>")
        write!(c, extras)
     end
     new(name, f, properties, extras, tag)::Component
end

The Interface portion of this module is actually built as a Toolips extension itself. Anyway, as you can see, the function f is provided. This is the one consistent field every servable must have. In that field you are able to write to the document with text how you normally would. That being said, Servable extensions can be used simply to generate one portion of your website while holding some information in a constructor. As soon as it is created, it is immediately dispatched to methods like write!, etc. Here is another, more simple example where we write a header.

import Toolips: Servable
mutable struct MyHeader <: Servable
    f::Function
    cs::Vector{Servable}
    function MyHeader(name = "Hello World")
        anim = Animation("fade_in")
        div_s = Style("div.myheaderstyle", color = "lightblue")
        header_div = divider("header_div", align = "center")
        heading = h(1, "Hello, welcome!", align = "center")
        style!(heading, "color" => "white")
        push!(header_div, heading)
        animate!(div_s, anim)
        cs = components(div_s, header_div)
        f(c::Connection) = write!(c, cs)
        new(f, cs)
    end
end

Is this the best way to serve your websites? It could be. This is somewhat of a manipulation to the paradigm of Toolips, but also allows for the creation of some other extensions.

Toolips.Servable — Type

abstract type Servable

Servables can be written to a Connection via thier f() function and the interface. They can also be indexed with strings or symbols to change properties

Consistencies

f::Function - Function whose output to be written to http. Must take a single

positonal argument of type ::Connection or ::AbstractConnection

creating connection extensions

Toolips.AbstractConnection — Type

abstract type AbstractConnection

Connections are passed through function routes and can have Servables written to it.

Consistencies

routes::Dict - A {String, Function} dictionary that the server references to

direct incoming connections.

http::Any - Usually an HTTP.Stream, however can be anything that is binded to

the Base.write method.

extensions::Dict - A {Symbol, ServerExtension} dictionary that can be used to

access ServerExtensions.

creating server extensions

Toolips.ServerExtension — Type

abstract type ServerExtension

Server extensions are loaded into the server on startup, and can have a few different abilities according to their type field's value. This value can be either a Symbol or a Vector of Symbols.

Consistencies

type::T where T == Vector{Symbol} || T == Symbol. The type can be :routing,

:func, :connection, or any combination inside of a Vector{Symbol}. :routing ServerExtensions must have an f() function that takes two dictionaries; e.g. f(r::Dict{String, Function}, e::Dict{Symbol, ServerExtension}) The first Dict is the dictionary of routes, the second is the dictionary of server extensions. :func server extensions will be ran everytime the server is routed. They will need to have the same f function, but taking a single argument as a connection. Lastly, :connection extensions are simply pushed to the connection.

toolips internals

If you're looking at the internals, you are probably good enough at reading documentation... Here are the doc-strings, my friend. Thank you for contributing.

Base.write — Method

Internals

write(s::SpoofStream, e::Any) -> _

A binding to Base.write that allows one to write to SpoofStream.text.

example

s = SpoofStream()
write(s, "hi")
println(s.text)
    hi

Base.write — Method

Internals

write(s::SpoofStream, e::Servable) -> _

A binding to Base.write that allows one to write a Servable to SpoofStream.text.

example

s = SpoofStream()
write(s, p("hello"))
println(s.text)
    <p id = "hello"></p>

Toolips.create_serverdeps — Function

Internals

create_serverdeps(name::String, inc::String) -> _

Creates the essential portions of the webapp file structure, where name is the project's name and inc is any extensions or strings to incorporate at the top of the file.

example

create_serverdeps("ToolipsApp")

Toolips.serverfuncdefs — Function

Core

serverfuncdefs(::AbstractVector, ::String, ::Integer,

::Dict) -> (::Function, ::Function, ::Function)

This method is used internally by a constructor to generate the functions add, start, and remove for the ServerTemplate.

example

Toolips._start — Function

Core - Internals

_start(routes::AbstractVector, ip::String, port::Integer,

extensions::Dict, c::Type) -> ::WebServer

This is an internal function for the ServerTemplate. This function is binded to the ServerTemplate.start field.

example

st = ServerTemplate()
st.start()

Toolips.generate_router — Function

Core - Internals

generate_router(routes::AbstractVector, server::Any, extensions::Dict,

        conn::Type)

This method is used internally by the _start method. It returns a closure function that both routes and calls functions.

example

server = Sockets.listen(Sockets.InetAddr(parse(IPAddr, ip), port))
if has_extension(extensions, Logger)
    extensions[Logger].log(1,
     "Toolips Server starting on port " * string(port))
end
routefunc, rdct, extensions = generate_router(routes, server, extensions,
                                                Connection)
@async HTTP.listen(routefunc, ip, port, server = server)

Toolips._log — Function

Extensions

_log(level::Int64, message::String, levels::Dict, out::String) -> _

Binded call for the field log() inside of Logger(). See ?(Logger) for more details on the field log. All arguments are fields of that type. Return is a printout into the REPL as well as an append to the log file, provided by the out URI. –––––––––

example (Closure from Logger)

log(level::Int64, message::String) = _log(level, message, levels, out)
log(message::String) = _log(1, message, levels, out)

Extensions

_log(http::HTTP.Stream, message::String) -> _

Binded call for the field log() inside of Logger(). This will log both to the JavaScript/HTML console. –––––––––

example (Closure from Logger)

log(http::HTTP.Stream, message::String) = _log(http, message)

Base.string — Function

Internals

string(r::Vector{UInt8}) -> ::String

Turns a vector of UInt8s into a string.

Interface

string(c::Component) -> ::String

Shows c as a string representation of itself.

example

c = divider("example", align = "center")
string(c)
    "divider: align = center"

Toolips.SpoofConnection — Type

SpoofConnection <: AbstractConnection

routes::Dict

http::SpoofStream

extensions::Dict Builds a fake connection with a SpoofStream. Useful if you want to write

a Servable without a server.

example

fakec = SpoofConnection()
servable = Component()
# write!(::AbstractConnection, ::Servable):
write!(fakec, servable)

field info

routes::Dict - A dictionary of routes, usually left empty.

http::SpoofStream - A fake http stream that instead writes output to a string.

extensions::Dict - A dictionary of extensions, usually empty.

constructors

SpoofStream(r::Dict, http::SpoofStream, extensions::Dict)

SpoofStream()

Toolips.SpoofStream — Type

SpoofStream

text::String The SpoofStream allows us to fake a connection by building a SpoofConnection

which will write to the SpoofStream.text field whenever write! is called. This is useful for testing, or just writing servables into a string.

example

stream = SpoofStream()
write(stream, "hello!")
println(stream.text)

    hello!
conn = SpoofConnection()
servab = Component()
write!(conn, servab)

field info

text::String - The text written to the stream.

constructors

SpoofStream()

Toolips.route_from_dir — Function

Extensions

routefromdir(dir::String) -> ::Vector{String}

Recursively appends filenames for a directory AND all subsequent directories.

example

x::Vector{String} = route_from_dir("mypath")

Base.show — Method

Interface

show(t::Base.TTY, x::Component) -> _

Shows a component as markdown in a terminal.

example

# In the terminal, elsewhere the component will show as HTML.
show(x)

Base.show — Method

Interface

show(x::Component) -> _

Shows a component as HTML.

example

show(x)

Toolips.show_log — Function

Extensions

show_log(level::Int64, message::String, levels::Dict{Any, Crayon},

            prefix::String, time::Any)

Prints a log to the screen.

example

show_log(1, "hello!", levels, "toolips> ", now()

[2022:05:23:22:01] toolips> hello!

Toolips.@L_str — Macro

Interface

L_str(s::String) -> ::String

Creates a literal string

example

x = 5
L"dollar_signx" # pretend dollar_sign is a dollar sign.

Toolips.has_extension — Method

Internals

has_extension(d::Dict, t::Type) -> ::Bool

Checks if d has an extension of type t.

example

if has_extension(d, Logger)
    d[:Logger].log("it has a logger, I think.")
end

Toolips.argsplit — Function

Internals

argsplit(args::Vector{AbstractString}) -> ::Dict{Symbol, Any}

Used by the getargs method to parse GET arguments into a Dict.

example

argsplit(["c=5", "b=8"])
    Dict(:c => 5, :b => 8)

Base.string — Method

Internals

string(r::Vector{UInt8}) -> ::String

Turns a vector of UInt8s into a string.

Toolips.showchildren — Function

Internals

showchildren(x::Component) -> ::String

Get the children of x as a markdown string.

example

c = divider("example")
child = p("mychild")
push!(c, child)
s = showchildren(c)
println(s)
"##### children
|-- mychild